+++ client/src/app/products/layout.tsx
@@ -7,10 +7,10 @@ import { Fragment, ReactNode } from "react"
 export default function ProductLayout({ children }: { children: ReactNode }) {
   return (
     <Fragment>
-      <Show when={!!localStorage.getItem("apiKey")}>
+      <Show when={!!localStorage.getItem("clientId")}>
         {children}
       </Show>
-      <Show when={!localStorage.getItem("apiKey")}>
+      <Show when={!localStorage.getItem("clientId")}>
         <div className="box">
           <p className="branding">Hold On</p>
           <p className="muted-text">Seems like you are not having an active subscription in this workspace to use/view this Product</p>