+++ src/api/products/fabric/dto/view-db-credential.dto.ts
@@ -1,4 +1,4 @@
-export class ViewDbApiKeyDto {
+export class ViewDbCredentialDto {
   dbId: string
   dbPassword: string
 }