+++ src/api/products/nftstudio/nftstudio.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Post, BadRequestException, Body } from "@nestjs/common"
 import { NftstudioService } from "./nftstudio.service"
 import { statusMessages } from "../../../constants/status-messages"
-import { ApiKeyAuthorizer, ApiKeyAuthorizerResponse } from "src/authorization/apikey-authorizer.decorator"
+import { CredentialAuthorizer, CredentialAuthorizerResponse } from "src/authorization/credential-authorizer.decorator"
 import { TokenAuthorizer, TokenAuthorizerResponse } from "src/authorization/token-authorizer.decorator"
 
 @Controller("products/nftstudio")
@@ -9,9 +9,9 @@ export class NftstudioController {
   constructor(private readonly nftstudioService: NftstudioService) { }
 
   @Post("createtx")
-  async createTransaction(@ApiKeyAuthorizer() ufak: ApiKeyAuthorizerResponse) {
+  async createTransaction(@CredentialAuthorizer() ufc: CredentialAuthorizerResponse) {
     try {
-      await this.nftstudioService.createTransaction(ufak.workspaceId)
+      await this.nftstudioService.createTransaction(ufc.workspaceId)
       return true
     }