+++ src/api/products/pay/pay.controller.ts
@@ -1,16 +1,16 @@
 import { Controller, Post, BadRequestException, Body } from "@nestjs/common"
 import { PayService } from "./pay.service"
 import { statusMessages } from "../../../constants/status-messages"
-import { ApiKeyAuthorizer, ApiKeyAuthorizerResponse } from "src/authorization/apikey-authorizer.decorator"
+import { CredentialAuthorizer, CredentialAuthorizerResponse } from "src/authorization/credential-authorizer.decorator"
 
 @Controller("products/pay")
 export class PayController {
   constructor(private readonly payService: PayService) { }
 
   @Post("createtx")
-  async createTransaction(@ApiKeyAuthorizer() ufak: ApiKeyAuthorizerResponse) {
+  async createTransaction(@CredentialAuthorizer() ufc: CredentialAuthorizerResponse) {
     try {
-      const transaction = await this.payService.createTransaction(ufak.workspaceId)
+      const transaction = await this.payService.createTransaction(ufc.workspaceId)
       return transaction
     }