+++ src/api/products/swap/swap.controller.ts
@@ -1,6 +1,6 @@
 import { BadRequestException, Controller, Post, Body } from "@nestjs/common"
 import { SwapService } from "./swap.service"
-import { ApiKeyAuthorizer, ApiKeyAuthorizerResponse } from "src/authorization/apikey-authorizer.decorator"
+import { CredentialAuthorizer, CredentialAuthorizerResponse } from "src/authorization/credential-authorizer.decorator"
 import { statusMessages } from "src/constants/status-messages"
 import { TokenAuthorizer, TokenAuthorizerResponse } from "src/authorization/token-authorizer.decorator"
 import { SearchTokensDto } from "./dto/search-tokens.dto"
@@ -23,9 +23,9 @@ export class SwapController {
   }
 
   @Post("createtx")
-  async createTransaction(@ApiKeyAuthorizer() ufak: ApiKeyAuthorizerResponse) {
+  async createTransaction(@CredentialAuthorizer() ufc: CredentialAuthorizerResponse) {
     try {
-      const transaction = await this.swapService.createTransaction(ufak.workspaceId)
+      const transaction = await this.swapService.createTransaction(ufc.workspaceId)
       return transaction
     }